% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashCard.R
\name{flashCard}
\alias{flashCard}
\title{Create a flash card}
\usage{
flashCard(
  data,
  frontColor = "#090e87",
  backColor = "#3443c9",
  front_text_color = "white",
  back_text_color = "white",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{df with front and back columns}

\item{frontColor}{frond side background color}

\item{backColor}{back side background color}

\item{front_text_color}{frond text color}

\item{back_text_color}{back text color}

\item{width}{width}

\item{height}{height}

\item{elementId}{element id}
}
\description{
Create a flash card  with desired data frame
}
\examples{
df1 <- data.frame(
front = c("Title front","contentfront", "content second line"),
back = c("Title back","content back", "second line")
)
flashCard(df1, elementId = "card", front_text_color = "grey")

}
